+++ lib/src/player_with_controls.dart
@@ -8,65 +8,63 @@ import 'package:flutter/material.dart';
 import 'package:video_player/video_player.dart';
 
 class PlayerWithControls extends StatelessWidget {
-  PlayerWithControls({Key key}) : super(key: key);
+  const PlayerWithControls({Key key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
     final ChewieController chewieController = ChewieController.of(context);
 
-    return Center(
-      child: Container(
-        width: MediaQuery.of(context).size.width,
-        child: AspectRatio(
-          aspectRatio:
-              chewieController.aspectRatio ?? _calculateAspectRatio(context),
-          child: _buildPlayerWithControls(chewieController, context),
-        ),
-      ),
-    );
-  }
+    double _calculateAspectRatio(BuildContext context) {
+      final size = MediaQuery.of(context).size;
+      final width = size.width;
+      final height = size.height;
 
-  Container _buildPlayerWithControls(
-      ChewieController chewieController, BuildContext context) {
-    return Container(
-      child: Stack(
+      return width > height ? width / height : height / width;
+    }
+
+    Widget _buildControls(
+      BuildContext context,
+      ChewieController chewieController,
+    ) {
+      final controls = Theme.of(context).platform == TargetPlatform.android
+          ? const MaterialControls()
+          : const CupertinoControls(
+              backgroundColor: Color.fromRGBO(41, 41, 41, 0.7),
+              iconColor: Color.fromARGB(255, 200, 200, 200),
+            );
+      return chewieController.showControls ? chewieController.customControls ?? controls : Container();
+    }
+
+    Stack _buildPlayerWithControls(ChewieController chewieController, BuildContext context) {
+      return Stack(
         children: <Widget>[
           chewieController.placeholder ?? Container(),
           Center(
             child: AspectRatio(
-              aspectRatio: chewieController.aspectRatio ??
-                  _calculateAspectRatio(context),
+              aspectRatio: chewieController.aspectRatio ?? chewieController.videoPlayerController.value.aspectRatio,
               child: VideoPlayer(chewieController.videoPlayerController),
             ),
           ),
           chewieController.overlay ?? Container(),
-          _buildControls(context, chewieController),
+          if (!chewieController.isFullScreen)
+            _buildControls(context, chewieController)
+          else
+            SafeArea(
+              child: _buildControls(context, chewieController),
+            ),
         ],
+      );
+    }
+
+    return Center(
+      child: SizedBox(
+        height: MediaQuery.of(context).size.height,
+        width: MediaQuery.of(context).size.width,
+        child: AspectRatio(
+          aspectRatio: _calculateAspectRatio(context),
+          child: _buildPlayerWithControls(chewieController, context),
+        ),
       ),
     );
-  }
-
-  Widget _buildControls(
-    BuildContext context,
-    ChewieController chewieController,
-  ) {
-    return chewieController.showControls
-        ? chewieController.customControls != null
-            ? chewieController.customControls
-            : Theme.of(context).platform == TargetPlatform.android
-                ? MaterialControls()
-                : CupertinoControls(
-                    backgroundColor: Color.fromRGBO(41, 41, 41, 0.7),
-                    iconColor: Color.fromARGB(255, 200, 200, 200),
-                  )
-        : Container();
-  }
-
-  double _calculateAspectRatio(BuildContext context) {
-    final size = MediaQuery.of(context).size;
-    final width = size.width;
-    final height = size.height;
-
-    return width > height ? width / height : height / width;
   }
 }