+++ lib/src/utils.dart
@@ -4,19 +4,28 @@ String formatDuration(Duration position) {
   int seconds = ms ~/ 1000;
   final int hours = seconds ~/ 3600;
   seconds = seconds % 3600;
-  var minutes = seconds ~/ 60;
+  final minutes = seconds ~/ 60;
   seconds = seconds % 60;
 
-  final hoursString = hours >= 10 ? '$hours' : hours == 0 ? '00' : '0$hours';
+  final hoursString = hours >= 10
+      ? '$hours'
+      : hours == 0
+          ? '00'
+          : '0$hours';
 
-  final minutesString =
-      minutes >= 10 ? '$minutes' : minutes == 0 ? '00' : '0$minutes';
+  final minutesString = minutes >= 10
+      ? '$minutes'
+      : minutes == 0
+          ? '00'
+          : '0$minutes';
 
-  final secondsString =
-      seconds >= 10 ? '$seconds' : seconds == 0 ? '00' : '0$seconds';
+  final secondsString = seconds >= 10
+      ? '$seconds'
+      : seconds == 0
+          ? '00'
+          : '0$seconds';
 
-  final formattedTime =
-      '${hoursString == '00' ? '' : hoursString + ':'}$minutesString:$secondsString';
+  final formattedTime = '${hoursString == '00' ? '' : '$hoursString:'}$minutesString:$secondsString';
 
   return formattedTime;
 }